+++ libs/ui/src/components/molecules/AddressDisplay/index.ts
@@ -1,1 +1,2 @@
-export * from './AddressDisplay';
+export { AddressDisplay } from './AddressDisplay';
+export type { AddressDisplayProps } from './AddressDisplay.types';
