+++ libs/ui/src/components/molecules/Banner/Banner.types.ts
@@ -1,0 +1,4 @@
+export type BannerProps = {
+  bannerText?: string;
+  className?: string;
+};
