+++ libs/ui/src/components/molecules/Banner/index.ts
@@ -1,1 +1,2 @@
-export * from './Banner';
+export { Banner } from './Banner';
+export type { BannerProps } from './Banner.types';
