+++ libs/ui/src/components/molecules/CSInput/index.ts
@@ -1,1 +1,1 @@
-export * from './CSInput';
+export { CSInput } from './CSInput';
