+++ libs/ui/src/components/molecules/CheckGate/CheckGate.types.ts
@@ -1,0 +1,7 @@
+import React, { ReactNode } from 'react';
+
+export type CheckGateProps = {
+  gateLabel: string;
+  fields: Array<ReactNode>;
+  onUnchecked?: () => void;
+};
