+++ libs/ui/src/components/molecules/CollapsibleCard/CollapsibleCard.types.ts
@@ -1,0 +1,7 @@
+import { CollapsibleProps } from '@radix-ui/react-collapsible';
+
+export type CollapsibleCardProps = CollapsibleProps & {
+  collapsibleContent?: React.ReactNode;
+  triggerLabel: string;
+  width: string;
+};
