+++ libs/ui/src/components/molecules/DataIndicator/DataIndicator.types.ts
@@ -1,0 +1,6 @@
+export type DataIndicatorProps = {
+  label?: string;
+  data?: string | number;
+  size?: 'lg' | 'sm';
+  info?: string;
+};
