+++ libs/ui/src/components/molecules/DataIndicator/index.ts
@@ -1,1 +1,2 @@
-export * from './DataIndicator';
+export { DataIndicator } from './DataIndicator';
+export type { DataIndicatorProps } from './DataIndicator.types';
