+++ libs/ui/src/components/molecules/Dialog/Dialog.tsx
@@ -1,7 +1,7 @@
 import React, { RefObject } from 'react';
 import { RiCloseFill } from 'react-icons/ri';
 
-import { DialogProps } from '../../../types/dialog.types';
+import { DialogProps } from './Dialog.types';
 import { Button, H5 } from '../../atoms';
 import {
   DialogRoot,
@@ -18,6 +18,7 @@ import {
   CloseIcon,
 } from './Dialog.styles';
 
+// TODO Fix this ref
 type Ref =
   | ((instance: HTMLDivElement | null) => void)
   | RefObject<HTMLDivElement>
