+++ libs/ui/src/components/molecules/Dialog/index.ts
@@ -1,1 +1,2 @@
-export * from './Dialog';
+export { Dialog, DialogTrigger, DialogContent } from './Dialog';
+export type { DialogProps } from './Dialog.types';
