+++ libs/ui/src/components/molecules/FieldAlert/FieldAlert.tsx
@@ -1,25 +1,17 @@
-import { forwardRef, RefObject } from 'react';
+import React from 'react';
+
 import { ParXs } from '../../atoms/Typography';
 
+import { FieldAlertProps } from './FieldAlert.types';
 import { FieldAlertWrapper } from './FieldAlert.styles';
 
-export type FieldAlertProps = {
-  className?: string;
-  message: string;
+export const FieldAlert: React.FC<
+  React.PropsWithChildren & FieldAlertProps
+> = ({ className, message, children }) => {
+  return (
+    <FieldAlertWrapper className={className}>
+      <ParXs>{message}</ParXs>
+      {children}
+    </FieldAlertWrapper>
+  );
 };
-
-type Ref =
-  | ((instance: HTMLBaseElement | null) => void)
-  | RefObject<HTMLBaseElement>
-  | null
-  | undefined;
-
-export const FieldAlert: React.FC<React.PropsWithChildren & FieldAlertProps> =
-  forwardRef(({ className, message, children }, ref: Ref) => {
-    return (
-      <FieldAlertWrapper className={className}>
-        <ParXs>{message}</ParXs>
-        {children}
-      </FieldAlertWrapper>
-    );
-  });
