+++ libs/ui/src/components/molecules/FieldAlert/FieldAlert.types.ts
@@ -1,0 +1,4 @@
+export type FieldAlertProps = {
+  className?: string;
+  message: string;
+};
