+++ libs/ui/src/components/molecules/FieldWrapper/FieldWrapper.types.ts
@@ -1,0 +1,14 @@
+import { FieldError } from 'react-hook-form';
+
+import {
+  ErrorMessage,
+  WarningMessage,
+  SuccessMessage,
+} from '../../../types/formAndField';
+
+export type FieldWrapperHelperTextFactoryProps = {
+  error?: ErrorMessage | FieldError;
+  warning?: WarningMessage;
+  success?: SuccessMessage;
+  helperText?: string;
+};
