+++ libs/ui/src/components/molecules/FieldWrapper/index.ts
@@ -1,1 +1,1 @@
-export * from './FieldWrapper';
+export { FieldWrapper } from './FieldWrapper';
