+++ libs/ui/src/components/molecules/FileInput/index.ts
@@ -1,1 +1,1 @@
-export * from './FileInput';
+export { FileInput } from './FileInput';
