+++ libs/ui/src/components/molecules/HighlightInputText/HighlightInputText.types.ts
@@ -1,0 +1,6 @@
+export type HighlightInputTextProps = {
+  color?: string;
+  description: string;
+  highlightColor?: string;
+  highlightInputId?: string;
+};
