+++ libs/ui/src/components/molecules/HighlightInputText/index.ts
@@ -1,1 +1,2 @@
-export * from './HighlightInputText';
+export { HighlightInputText } from './HighlightInputText';
+export type { HighlightInputTextProps } from './HighlightInputText.types';
