+++ libs/ui/src/components/molecules/ImageInput/index.ts
@@ -1,1 +1,1 @@
-export * from './ImageInput';
+export { ImageInput } from './ImageInput';
