+++ libs/ui/src/components/molecules/InputSelect/index.ts
@@ -1,1 +1,1 @@
-export * from './InputSelect';
+export { InputSelect } from './InputSelect';
