+++ libs/ui/src/components/molecules/ProfileAvatar/ProfileAvatar.types.ts
@@ -1,0 +1,6 @@
+import { Avatar } from '../../atoms';
+
+export type ProfileAvatarProps = Parameters<typeof Avatar>[0] & {
+  address?: string;
+  image?: string;
+};
