+++ libs/ui/src/components/molecules/TimePicker/TimePicker.types.ts
@@ -1,0 +1,8 @@
+import { Field, OptionType } from '../../../types';
+
+export type TimePickerProps = Field & {
+  defaultValue?: string;
+  options?: OptionType[];
+  selectId?: string;
+  selectPlaceholder?: string;
+};
