+++ libs/ui/src/components/molecules/TimePicker/index.ts
@@ -1,1 +1,2 @@
-export * from './TimePicker';
+export { TimePicker } from './TimePicker';
+export type { TimePickerProps } from './TimePicker.types';
