+++ libs/ui/src/components/molecules/Toast/index.ts
@@ -1,1 +1,2 @@
-export * from './Toast';
+export { Toast } from './Toast';
+export { ToastProps } from './Toast.types';
