+++ libs/ui/src/components/molecules/WrappedCheckbox/index.ts
@@ -1,1 +1,1 @@
-export * from './WrappedCheckbox';
+export { WrappedCheckbox } from './WrappedCheckbox';
