+++ libs/ui/src/components/molecules/WrappedFileInput/index.ts
@@ -1,1 +1,1 @@
-export * from './WrappedFileInput';
+export { WrappedFileInput } from './WrappedFileInput';
