+++ libs/ui/src/components/molecules/WrappedInputSelect/index.ts
@@ -1,1 +1,1 @@
-export * from './WrappedInputSelect';
+export { WrappedInputSelect } from './WrappedInputSelect';
