+++ libs/ui/src/components/molecules/WrappedRadio/index.ts
@@ -1,1 +1,1 @@
-export * from './WrappedRadio';
+export { WrappedRadio } from './WrappedRadio';
