+++ libs/ui/src/components/molecules/WrappedTextArea/index.ts
@@ -1,1 +1,1 @@
-export * from './WrappedTextArea';
+export { WrappedTextArea } from './WrappedTextArea';
