+++ libs/ui/src/components/organisms/AppSwitcher/AppSwitcher.tsx
@@ -1,32 +1,12 @@
 import React from 'react';
 import { RiArrowDropDownLine } from 'react-icons/ri';
-import { IconType } from 'react-icons';
 
-import { Dropdown, DropdownMenuItem } from '../../molecules';
-import { AppSwitcherLink, AppSwitcherTrigger } from './AppSwitcher.styles';
-import { useBreakpoint } from '../../../hooks/useMediaQuery';
 import { widthQuery } from '../../../theme/global/breakpoints';
+import { useBreakpoint } from '../../../hooks/useMediaQuery';
+import { Dropdown, DropdownMenuItem } from '../../molecules';
 
-interface IApp {
-  name: string;
-  url: string;
-  icon?:
-    | IconType
-    | React.FunctionComponent<
-        React.SVGProps<SVGSVGElement> & {
-          title?: string | undefined;
-        }
-      >;
-}
-
-export type AppSwitcherProps = {
-  trigger: IApp;
-  apps: IApp[];
-  spacing?: string;
-  width?: string;
-  menuBg?: string;
-  className?: string;
-};
+import { IApp, AppSwitcherProps } from './AppSwitcher.types';
+import { AppSwitcherLink, AppSwitcherTrigger } from './AppSwitcher.styles';
 
 function getDropdownApps(apps: IApp[]) {
   return apps.map((app, index) => (
