+++ libs/ui/src/components/organisms/AppSwitcher/AppSwitcher.types.ts
@@ -1,0 +1,16 @@
+import { IconType } from 'react-icons';
+
+export interface IApp {
+  name: string;
+  url: string;
+  icon?: IconType | React.FunctionComponent<React.SVGProps<SVGSVGElement>>;
+}
+
+export type AppSwitcherProps = {
+  trigger: IApp;
+  apps: IApp[];
+  spacing?: string;
+  width?: string;
+  menuBg?: string;
+  className?: string;
+};
